+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/assets/application/base/service/AssetBaseService.java
@@ -47,6 +47,8 @@ public interface AssetBaseService {
 
     AssetBase updateQualityType(String assetId, QualityType qualityType);
 
+    List<String> getDistinctFilterValues(String fieldName, String startWith, Integer size, Owner owner, List<String> inAssetIds);
+
     List<String> getSearchableValues(String fieldName, String startWith, Integer size, Owner owner, List<String> inAssetIds);
 
     List<String> getAssetIdsInImportState(ImportState... importStates);