+++ CHANGELOG.md
@@ -6,7 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 _**For better traceability add the corresponding GitHub issue number in each changelog entry, please.**_
-## [UNRELEASED - DD.MM.YYYY]
+## [UNRELEASED -09.09.2024]
+
+### Added
+- #43 Implemented backend API for search filtering (asBuilt and asPlanned).
+- #43 Integrated search functionality in the CFX frontend.
+- #43 Added integration tests for search functionality based on legacy repo code.
+### Fixed
+- #43 Resolved missing API issues causing search functionality in CFX frontend to fail.
 
 ## [13.0.2 - 16.08.2024]